Projected annual saving

+++ python_introduction/finance_calculator.py
@@ -1,6 +1,6 @@
 # Personal Finance Calculator
-income = int(input("Enter your monthly income: "))
-expenses = int(input("Enter your total monthly expenses: "))
+income = float(input("Enter your monthly income: "))
+expenses = float(input("Enter your total monthly expenses: "))
 savings = income - expenses
 Savings = savings * 12 + (savings * 12 * 0.05)
 print("Your monthly savings are ${savings}.")
